backend/pyproject.toml
[project]
name = "agent-workflow-backend"
version = "0.1.0"
description = "Backend for the Agent Workflow System"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "apscheduler",
    "claude-code-sdk",
    "structlog",
    "pydantic",
    "python-multipart",
    "pyyaml",
    "jinja2",
    "httpx",
    "websockets",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
